Fix instructions that pointed at things nobody can run
Every one of these sends a curator somewhere that does not exist, which is
the same defect in four places.

THE PRODUCTS EMPTY SCREEN told her to run `npm run import:products`, and
products/actions.ts claimed "a CSV importer already exists
(scripts/import-products.mjs)". It does not exist and is not in
package.json. The one instruction on an empty screen was a dead end.

Both now say there is no bulk importer, and the comment records what
should happen instead: when bulk import is genuinely needed — and
data/decor-candidates.json holds 80 sourced rows waiting for exactly that
— it belongs as a DESK SCREEN rather than a script, because the database
is unreachable from any laptop and a script nobody can run is worse than
no script at all.

THE COUNT COLUMNS were a real visible bug, not just a type lie. count(*)
is bigint and node-postgres returns bigint as a STRING, so a column typed
`number` survives `drafts > 0` on coercion but fails `drafts === 1` for
"1" — the plural under it read "1 drafts" for every destination. Six
counts cast to ::int in the query, which makes the declared type true
rather than papering over it.

THE COVERAGE BOARD pointed its Games column at the connections matrix
because no games list existed. One does now.

THE --overwrite ASYMMETRY is documented in the three seeders that lack it.
menus and drinks can let the file beat the curator; games, dishes and
destinations cannot, so a desk edit there is unrevertible by any part of
the pipeline. Defensible, but five scripts describe themselves in
near-identical words and only two have the escape hatch.

+++ src/app/desk/(signed-in)/destinations/[id]/page.tsx
@@ -51,16 +51,16 @@ export default async function DestinationPage({
     }>(
       `select v.id, v.version, v.status::text as status, v.authored_by, v.note,
               v.published_at,
-              (select count(*) from revelle r where r.voice_id = v.id) as issued
+              (select count(*)::int from revelle r where r.voice_id = v.id) as issued
          from world_voice v
         where v.world_id = $1
         order by v.version desc`,
       [id]
     ),
     queryOne<{ revelles: number; menus: number; sections: number }>(
-      `select (select count(*) from revelle where world_id = $1) as revelles,
-              (select count(*) from menu_world where world_id = $1) as menus,
-              (select count(*) from world_section where world_id = $1) as sections`,
+      `select (select count(*)::int from revelle where world_id = $1) as revelles,
+              (select count(*)::int from menu_world where world_id = $1) as menus,
+              (select count(*)::int from world_section where world_id = $1) as sections`,
       [id]
     ),
   ]);
